\name{pagerank.ddR-package}
\alias{pagerank.ddR-package}
\alias{pagerank.ddR}
\docType{package}
\title{Distributed pagerang for Big Data}
\description{
    \pkg{pagerank.ddR} provides a distributed function for pagerank 
}

\details{
    \tabular{ll}{
        Package: \tab pagerank.ddR\cr
        Type: \tab Package\cr
        Version: \tab 0.1.0\cr
        Date: \tab 2015-11-11\cr
    }
 Main Functions:
 \itemize{
  \item {dpagerank:} {It is a distributed version of pagerank.}
 }
}

\author{
Arash Fard, Vishrut Gupta
}

\keyword{pagerank}
\keyword{distributed pagerank}

